(MODULES-6094) Prefer puppet-agent binaries
This commit updates the puppet_conf task to prefer `puppet` and `puppet.bat`
binaries installed by the puppet-agent package over searching `$PATH` for
puppet. This change allows the task to be executed in node environments where
the puppet-agent bindir is not on the default path. The logic implemented has
been copied from the facter_task module.